website/requirements.txt
Flask
Flask-Login
Flask-Script
pycrypto
